fix(anolisa): split qoder settings helpers
- Move qoder settings merge and prune helpers into a submodule.
- Keep driver lifecycle code focused on CLI and receipt handling.
